documentation/requirements.txt
python-barcode
pillow
pycups
reportlab 
pycups